requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
sqlalchemy==2.0.35
python-multipart==0.0.12
pillow==11.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==24.1.0
jinja2==3.1.4
